Fix Infused Seeds crash
getDrops() can't actually determine the drops of the infused crops, as it is called after the tile entity ceases to exist. breakBlock() is called at the proper time, but having the code there doesn't play well with the EnderIO planter, which is why the code was moved to getDrops() in the first place.